requirements.txt
dj-database-url==0.5.0
Django==2.2.1
django-dajax==0.9.2
django-dajaxice==0.7
django-dbsettings==0.10.0
django-heroku==0.3.1
django-synchro==0.7
djangorestframework==3.9.4
gunicorn==19.9.0
psycopg2==2.8.3
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
whitenoise==4.1.2